pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tollbooth-authority"
version = "0.1.0"
description = "Tollbooth Authority — Certified Purchase Order Service"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "Lonnie VanZandt" }]

dependencies = [
    "fastmcp>=2.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "PyJWT>=2.8.0",
    "cryptography>=42.0.0",
    "tollbooth-dpyc>=0.1.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/tollbooth_authority"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
